More detailed Instructions
added more details to the instruction page so the player has a better understanding of how to play the game and what to do during gameplay.
I also tried to fix the walls. For some reason the player-object sometimes can't move left (or sometimes right) horizontally when it is directly against a horizontal wall.